+++ examples/aspnet.mvc/Content/libs/js/ng-upload.js
@@ -1,4 +1,4 @@
-// Version 0.2.0 
+// Version 0.3.1
 // AngularJS simple file upload directive
 // this directive uses an iframe as a target
 // to enable the uploading of files without
@@ -6,111 +6,155 @@
 //
 // <div ng-app="app">
 //   <div ng-controller="mainCtrl">
-//    <form action="/uploads" ng-upload="results()"> 
+//    <form action="/uploads" ng-upload> 
 //      <input type="file" name="avatar"></input>
-//      <input type="submit" value="Upload"></input>
+//      <input type="submit" value="Upload" 
+//         upload-submit="submited(content, completed)"></input>
 //    </form>
 //  </div>
 // </div>
 //
 //  angular.module('app', ['ngUpload'])
 //    .controller('mainCtrl', function($scope) {
-//      $scope.results = function(content) {
-//        console.log(content);
+//      $scope.submited = function(content, completed) {
+//        if (completed) {
+//          console.log(content);
+//        }
 //      }  
 //  });
 //
 angular.module('ngUpload', [])
-  .directive('ngUpload', function () {
+  .directive('uploadSubmit', ['$parse', function ($parse) {
       return {
-          restrict: 'A',
+          restrict: 'AC',
           link: function (scope, element, attrs) {
 
               // Options (just 1 for now) 
-              // Each option should be prefixed with 'upload-Options-' or 'uploadOptions'
+              // Each option should be prefixed with 'upload-options-' or 'uploadOptions'
               // {
               //    // specify whether to enable the submit button when uploading forms
               //    enableControls: bool 
               // }
               var options = {};
-              options.enableControls = attrs['uploadOptionsEnableControls'];
-              
-              // get scope function to execute on successful form upload
-              if (attrs['ngUpload']) {
+              options.enableControls = attrs.uploadOptionsEnableControls;
 
-                  element.attr("target", "upload_iframe");
-                  element.attr("method", "post");
+              // "ng:upload" is excluded as it fails some test Specs
+              var ngUploadSyntaxes = ["ng-upload", "ng_upload", "x-ng-upload", "data-ng-upload"];
 
-                  // Append a timestamp field to the url to prevent browser caching results
-                  element.attr("action", element.attr("action") + "?_t=" + new Date().getTime());
+              // Find an enclosing form
+              var form = null;
 
-                  element.attr("enctype", "multipart/form-data");
-                  element.attr("encoding", "multipart/form-data");
+              // First attempt using the attribute syntax, i.e. <form ngUploadSyntax />
+              angular.forEach(ngUploadSyntaxes,
+                  function (ngUploadSyntax) {
+                      // if a form has already been found, . . .
+                      if (form)
+                          return; // stop searching, jump out!
 
-                  // Retrieve the callback function
-                  var fn = attrs['ngUpload'].split('(')[0];
-                  var callbackFn = scope.$eval(fn);
-                  if (callbackFn == null || callbackFn == undefined || !angular.isFunction(callbackFn))
-                  {
-                      var message = "The expression on the ngUpload directive does not point to a valid function.";
-                      // console.error(message);
-                      throw message + "\n";
-                  }                      
+                      var forms = element.closest('form[' + ngUploadSyntax + ']');
 
-                  // Helper function to create new iframe for each form submission
-                  var addNewDisposableIframe = function (submitControl) {
-                      // create a new iframe
-                      var iframe = $("<iframe id='upload_iframe' name='upload_iframe' border='0' width='0' height='0' style='width: 0px; height: 0px; border: none; display: none' />");
+                      if (forms.length > 0) // If a form is found
+                          form = forms.first(); // jqery always return a collection, so pick first
+                  }
+              );
 
-                      // attach function to load event of the iframe
-                      iframe.bind('load', function () {
+              if (form == null) {
+                  // Second attempt using the class syntax, i.e. <form class='ngUploadSyntax' />
+                  angular.forEach(ngUploadSyntaxes,
+                      function (ngUploadSyntax) {
+                          // if a form has already been found, . . .
+                          if (form)
+                              return; // stop searching, jump out!
 
-                              // get content - requires jQuery
-                              var content = iframe.contents().find('body').text();
+                          var forms = element.closest('form.' + ngUploadSyntax);
 
-                              // execute the upload response function in the active scope
-                              scope.$apply(function () { callbackFn(content, content !== "" /* upload completed */); });
-
-                              // remove iframe
-                              if (content != "") // Fixes a bug in Google Chrome that dispose the iframe before content is ready.
-                                  setTimeout(function () { iframe.remove(); }, 250);
-
-                              //if (options.enableControls == null || !(options.enableControls.length >= 0))
-                              submitControl.attr('disabled', null);
-                              submitControl.attr('title', 'Click to start upload.');
-                          });
-
-                      // add the new iframe to application
-                      element.parent().append(iframe);
-                  };
-
-                  // 1) get the upload submit control(s) on the form (submitters must be decorated with the 'ng-upload-submit' class)
-                  // 2) attach a handler to the controls' click event
-                  $('.upload-submit', element).click(
-                      function () {
-
-                          addNewDisposableIframe($(this) /* pass the submit control */);
-
-                          scope.$apply(function () { callbackFn("Please wait...", false /* upload not completed */); });
-
-                          //console.log(angular.toJson(options));
-
-                          var enabled = true;
-                          if (options.enableControls === null || options.enableControls === undefined || options.enableControls.length >= 0) {
-                              // disable the submit control on click
-                              $(this).attr('disabled', 'disabled');
-                              enabled = false;
-                          }
-
-                          $(this).attr('title', (enabled ? '[ENABLED]: ' : '[DISABLED]: ') + 'Uploading, please wait...');
-
-                          // submit the form
-                          $(element).submit();
+                          if (forms.length > 0) // If a form is found
+                              form = forms.first(); // jqery always return a collection, so pick first
                       }
-                  ).attr('title', 'Click to start upload.');
+                  );
               }
-              else
-                  console.log("No callback function found on the ngUpload directive.");
+
+              if (form == null)
+                  // Complain
+                  throw "[uploadSubmit] Directive can only be used inside a form.";
+
+              if (!attrs.uploadSubmit && form.attr("ng-upload"))
+                  attrs.uploadSubmit = form.attr("ng-upload");
+
+              if (!attrs.uploadSubmit) {
+                  throw "[uploadSubmit] A valid callback function must be supplied.";
+              }
+
+              // Retrieve the callback function
+              var fnExpr = attrs.uploadSubmit.split('(')[0];
+              var fn = scope.$eval(fnExpr); //$parse(attrs.uploadSubmit);
+
+              if (!angular.isFunction(fn)) {
+                  var message = "[uploadSubmit] Expression does not point to a valid function.";
+                  throw message + "\n";
+              }
+
+              element.bind('click', function () {
+
+                  // new id is generated for the iframe as every click
+                  form.attr('target', new Date().getTime());
+
+                  // create a new iframe
+                  var iframe = angular.element("<iframe id='" + form.attr("target") + "' name='" + form.attr("target") + "' border='0' width='0' height='0' style='width: 0px; height: 0px; border: none; display: none' />");
+
+                  // attach function to load event of the iframe
+                  iframe.bind('load', function () {
+                      // get content - requires jQuery
+                      var content = iframe.contents().find('body').text();
+
+                      // execute the upload response function in the active scope
+                      scope.$apply(function () {
+                          fn(content, true);
+                      });
+
+                      // remove iframe
+                      if (content !== "") { // Fixes a bug in Google Chrome that dispose the iframe before content is ready.
+                          setTimeout(function () { iframe.remove(); }, 250);
+                          element.attr('disabled', null);
+                          element.attr('title', 'Click to start upload.');
+                      }
+                  });
+
+                  // add the new iframe to application
+                  form.parent().append(iframe);
+
+                  scope.$apply(function () {
+                      fn("Please wait...", false);
+                  });
+
+                  var enabled = true;
+                  if (!options.enableControls) {
+                      // disable the submit control on click
+                      element.attr('disabled', 'disabled');
+                      enabled = false;
+                  }
+                  // why do we need this???
+                  element.attr('title', (enabled ? '[ENABLED]: ' : '[DISABLED]: ') + 'Uploading, please wait...');
+
+                  // submit the form - requires jQuery
+                  form.submit();
+
+              }).attr('title', 'Click to start upload.');
           }
       };
-  });
+  }])
+  .directive('ngUpload', ['$parse', function ($parse) {
+      return {
+          restrict: 'AC',
+          link: function (scope, element, attrs) {
+              /*
+                  target iframe identity are autogenerated to avoid iframe 'Permission Denied' errors
+              */
+              // element.attr("target", "upload_iframe");
+              // element.attr('target', new Date().getTime());
+              element.attr("method", "post");
+              element.attr("enctype", "multipart/form-data");
+              element.attr("encoding", "multipart/form-data");
+          }
+      };
+  }]);