+++ testacular.conf.js
@@ -58,7 +58,7 @@ autoWatch = true;
 // - Safari (only Mac)
 // - PhantomJS
 // - IE (only Windows)
-browsers = ['Chrome'];
+browsers = ['IE', 'Firefox'];
 
 
 // If browser does not capture in given timeout [ms], kill it